requirements.txt
PySimpleGUI
python-docx
pypdf
python-pptx
openpyxl
openai
langchain
llama-cpp-python==0.2.90
faiss-cpu
tiktoken
watchdog
langchain-text-splitters
langchain-community
fastapi
uvicorn
python-multipart
requests
tqdm
